requirements.txt
feedparser>=6.0.0
google-genai>=0.5.0
Pillow>=10.0.0
requests>=2.31.0
